Add DC removal and baseband CORDIC frequency parameters
* Add DC removal parameter. By default is false so
devices that do not support it to remain unaffected
* Add baseband tuning frequency. This can be used
to digitally shift the center frequency. In many cases
this shift should be large enough and the LO Offset
parameter can drastically increase the required bandwidth
if the values are high enough. However, devices that
support the baseband tuning, can do this efficiently
in the FPGA, minimizing or even zero'ing the LO Offset
parameter value. This is an option for people that
know what they are doing. In some devices, setting
the baseband frequency, does not compensate the
RF center frequency and should be done manualy

Signed-off-by: Manolis Surligas <surligas@gmail.com>