Use 12 kHz IF for OGG recordings
This changes makes it easier and more likely to be able to recover packets from OGG recordings in the BPSK flowgraph, in the case where there is no suitable demodulator within gr-satnogs.
At the moment the flowgraph shifts the centre of the doppler-corrected receiver passband up to baud_rate Hz. This is fine if the signal is precisely on frequency, but if the signal is even slightly off-frequency, part of it is clipped off, making it un-decodable.
This change uses a 12 kHz IF for all recordings, and widens the filter bandwidth to make best use of the 24 kHz available in the 48 kHz OGG recording.

This allows recordings of 1k2 BPSK from the various FunCube transponders to be replayed back through Funcube Dashboard, and for other higher-rate BPSK recordings to be replayed through gr-satellites.
This change has been tested on station #232, using a range of satellite transmitters.

Signed-off-by: Mark Jessop <vk5qi@rfhead.net>